Revert the (ineffectual) anti-starvation change in rev 76556.
Summary: Turns out epoll_wait() starves the TCP sockets the same way memcached did, so there's no point returning control to it.

Reviewed By: marc

Test Plan: Passes memcached test suite (which it had better, since this reverts the input handling code back to the way it used to be)

Revert: This *is* a reversion, so if you revert this you'll really be un-reverting the original change, which didn't help.