fix race condition in do_add_delta
Summary: Even though we have the cache lock in do_add_delta, some other thread may be transmitting the data in the object.  So when the refcount is greater than 1, we need to allocate a new object, even if we can fit in the old one.

Reviewed By: marc

Test Plan: - test/test.py -a passes
           - 24 hours with the 3 clients running the memcached stress test have not failed.

Revert: OK