fixed arithmetic commands, modified stats gathering, added stress test
Summary: arithmetic commands are properly processed now.  reverted to the old way of determining whether replace-in-place can be handled.  the reason the more aggressive algorithm does not work is because the flags & length string is stored in ascii.  changing the length is not trivial.  with the storage refactoring, it is possible.
         
         two bytecount statistics are now gathered.  one is the amount of memory allocated (so key + value + ALL item storage overhead) and another statistic with just all key + value.  previously we were gathering key + value + sizeof(stritem).  that doesn't really make sense since that doesn't take into account all item storage overhead (such as the tail of a slab which can't fit an item).
         
         finally, I ported the stress test I was using on the storage refactoring.  it would have caught the bugs with the arithmetic ops (in fact, that's how I ended up catching it).

Reviewed By: marc

Test Plan: ran stress test for 15 minutes, saw no data corruption.

Revert: OK