fixes to item_walk, a unit test for item_walk
Summary: item_walk had some bugs for extracting the middle part of an item.  we never do that (we get the entire item for arithmetic/send/receive, and we get the end for memset/stamping), so this probably shouldn't actually make a difference.  still, it is good to fix.
         
         also added some unit tests for item_walk (which is how the bugs were found).
         
         This includes http://www.dev.facebook.com/intern/diffcamp/?tab=review&revisionID=13628

Reviewed By: marc

Test Plan: all flat allocator tests pass

Revert: OK