open the maps file before we drop root privileges
Summary: once we drop root privileges, we cannot open the maps file.

Reviewed By: ps

Test Plan: started memcached with sudo and -u nobody.  did a stats maps.  flooded memcached with sets to force it to do a brk.  did a stats maps.  diffed the outputs to verify that rewinding and rereading the file works.

Revert: OK